break up utils.py to cmds.py and io.py, function renames
utils.py to cmds.py and io.py, handle_exceptions to
translate_exceptions, json_echo to echo_json
